refactor: delimiter 클래스 분리
